Move config verification in separate class
